requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
langdetect==1.0.9
pyyaml==6.0.1
pytest==8.1.1
httpx==0.27.0
